requirements.txt
google-generativeai>=0.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
